pyproject.toml
[project]
name = "node-agent"
version = "0.1.0"
description = "Node Agent"
authors = [
    { name = "Jerome Rastit" }
]
dependencies = [
    "dynaconf",
    "dataclasses_jsonschema",
    "sqlalchemy",
    "fastapi",
    "uvicorn",
    "psycopg2",
    "psycopg2-binary",
    "decorators",
    "gunicorn",
    "grpcio",
    "ansible",
    "ansible-base",
    "supabase",
]
[project.optional-dependencies]
test = [
    "flake8",
    "pylint",
    "pylint-exit",
    "anybadge",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "alembic",
    "alembic-utils",
]
